Renamed loader module
